refactor(stereo): add descriptions and minor changes to stereo functions

+++ depthy/stereo/dissimilarity_measures.py
@@ -1,7 +1,5 @@
 import numpy as np
 
-from depthy.misc import Normalizer
-
 # sum of absolute differences
 sad = lambda window, kernel: np.sum(np.abs(np.subtract(window, kernel)))
 sad_vectorized = lambda window, kernel, axis=1: np.sum(np.abs(np.subtract(window, kernel)), axis=axis)
@@ -13,7 +11,7 @@ def xor_similarity(census_values, census):
     Dissimilarity measure based on XOR operation from previously computed census values.
     :param census_values:
     :param census:
-    :return: H x W x D array as numpy uint64 type.
+    :return: H x W x D array as numpy uint32 type.
     """
 
     xor_img = np.int64(np.bitwise_xor(np.int64(census_values), census))
